add not_if guards to make execute resources idempotent and subsequent chef runs faster

+++ recipes/gitlab.rb
@@ -65,6 +65,7 @@ execute "install_gitlab" do
   user node['gitlab']['user']
   command "/home/#{node['gitlab']['user']}/gitlab-shell/bin/install"
   action :run
+  not_if { File.exists?("/home/#{node['gitlab']['user']}/repositories") }
 end
 
 # Checkout gitlabhq
@@ -124,13 +125,15 @@ execute "install_gitlabhq" do
   cwd node['gitlab']['home']
   command "bundle install --deployment --without development test postgres"
   action :run
+  not_if { File.exists?("#{node['gitlab']['home']}/vendor/bundle") }
 end
 
 execute "setup_gitlabhq" do
   user node['gitlab']['user']
   cwd node['gitlab']['home']
-  command "bundle exec rake gitlab:setup RAILS_ENV=production force=yes"
+  command "bundle exec rake gitlab:setup RAILS_ENV=production force=yes && touch .gitlab-setup"
   action :run
+  not_if { File.exists?("#{node['gitlab']['home']}/.gitlab-setup") }
 end
 
 # cp lib/support/init.d/gitlab /etc/init.d/gitlab
